Refactor library-service

+++ library-service/src/main/java/com/libraryapi/libraryservice/dto/BookRequest.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BookRequest {
-    private int bookId;
+    private long bookId;
 }
