+++ library-service/src/main/java/com/libraryapi/libraryservice/dto/LibraryBookDto.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 @Data
 public class LibraryBookDto {
-    private int bookId;
+    private long bookId;
     private LocalDateTime takenAt;
     private LocalDateTime returnDueAt;
